Update donut chart per recent changes
